requirements.txt
google-generativeai
python-dotenv
pydantic
guardrails-ai
langchain
langchain-google-genai